refactor(super-admin): SUPER_ADMIN role, service and controller

+++ src/main/java/com/GesCom/repository/PlanSuscripcionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PlanSuscripcionRepository extends JpaRepository<PlanSuscripcion, Integer> {
+public interface PlanSuscripcionRepository extends JpaRepository<PlanSuscripcion, Long> {
 
     Optional<PlanSuscripcion> findByNombre(String nombre);
 }
